Update viewModel, move it to feature module, some small changes.
